improve projection graph and breakdown of expenses
In this commit the projection graph was modified to include the users projected income increase cumulative
savings. The user is now able to tell at a glance how inflation and salary are going to affect their
savings over the next five years.
We also modified the breakdown of projected expenses. This section now allows the user to see how the
price of their individual expenses are set to increase.